+++ R/acq-interactive-utils.R
@@ -693,7 +693,8 @@ set_session_name_interactive <- function(session.name = NULL) {
 #'
 #' @export
 #'
-f.trigger.message <- function(n = 1L) {
+f.trigger.message <- function(n = 1L, delay = 0) {
+  Sys.sleep(delay)
   n <- as.integer(n)
   if (n < 1L) return(TRUE)
   if (n == 1L) {